+++ scripts/macros.js
@@ -1,7 +1,11 @@
-import {riteOfAgony} from './macros/classFeatures/bloodHunter/rites/riteOfAgony.js'
-import {riteOfMarking} from './macros/classFeatures/bloodHunter/rites/riteOfMarking.js'
+import { riteOfAgony } from './macros/classFeatures/bloodHunter/rites/riteOfAgony.js';
+import { riteOfBlindness } from './macros/classFeatures/bloodHunter/rites/riteOfBlindness.js';
+import { riteOfMarking } from './macros/classFeatures/bloodHunter/rites/riteOfMarking.js';
+import { riteOfSiphoning } from './macros/classFeatures/bloodHunter/rites/riteOfSiphoning.js';
 
 export let macros = {
     'riteOfAgony': riteOfAgony,
-    'riteOfMarking': riteOfMarking
+    'riteOfBlindness': riteOfBlindness,
+    'riteOfMarking': riteOfMarking,
+    'riteOfSiphoning': riteOfSiphoning
 }