+++ scripts/module.js
@@ -2,6 +2,7 @@ import {bloodHunter} from './bloodHunter.js';
 import {constants} from './constants.js';
 import {helpers} from './helpers.js';
 import {macros} from './macros.js';
+import {queue} from './utility/queue.js';
 import {registerFeatures} from './customFeatures.js';
 import {remoteAimCrosshair, remoteDialog, remoteDocumentDialog, remoteDocumentsDialog, remoteMenu} from './utility/remoteDialog.js';
 export let socket;
@@ -12,6 +13,7 @@ Hooks.once('init', async function() {
 });
 
 Hooks.once('socketlib.ready', async function() {
+	console.log("5e Content | Registering sockets!");
 	socket = socketlib.registerModule('5e-content');
 	socket.register('remoteDialog', remoteDialog);
 	socket.register('remoteDocumentDialog', remoteDocumentDialog);
@@ -20,9 +22,15 @@ Hooks.once('socketlib.ready', async function() {
 	socket.register('remoteMenu', remoteMenu);
 });
 
+Hooks.once('ready', async function () {
+	console.log("5e Content | Registering Automation Hooks");
+	Hooks.on('midi-qol.preAttackRoll', macros.riteOfBlindness.reactionDefense);
+});
+
 globalThis['sorelliaAutomations'] = {
 	bloodHunter,
 	constants,
 	helpers,
+	queue,
 	macros
 }