+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 0.5.03 Release:
+The next phase of the blood hunter release is here! This version brings the remaining rites to automate down to just 3, as well as a few various class features. Also included are some new utility features which were implemented to aid in the creation of the automation for the rite of blindness, which is this module's first Reaction-based automation inclusion!
+
+## New Content:
+* Further automation of blood rite runes.
+
+## Bug Fixes
+* Streamlined several parts of the codebase in regards to blood rites to make them flow better and be less... hacky, in particular with vital sacrifices.
+
+## Update Notes
+* Rite of Blindness, Exsanguination and Siphoning have all been fully automated.
+* All Rites have had their codebase improved and made more efficient, should iron out any final kinks in them programmatically.
+
+## Update Notes:
+
 # 0.5.02 Release:
 
 ## New Content: